+++ src/app/app.module.ts
@@ -12,11 +12,11 @@ import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NbThemeModule } from '@nebular/theme';
 
-import { NbAppComponent } from './app.component';
-import { NbLayoutDirectionToggleComponent } from './layout-direction-toggle/layout-direction-toggle.component';
-import { NbDynamicToAddComponent } from '../playground/shared/dynamic.component';
-import { NbPlaygroundSharedModule } from '../playground/shared/shared.module';
-import { NbLayoutThemeToggleComponent } from './layout-theme-toggle/layout-theme-toggle.component';
+import { AppComponent } from './app.component';
+import { LayoutDirectionToggleComponent } from './layout-direction-toggle/layout-direction-toggle.component';
+import { LayoutThemeToggleComponent } from './layout-theme-toggle/layout-theme-toggle.component';
+import { ComponentsOverlayComponent } from './components-list/components-overlay.component';
+import { ComponentsListComponent} from './components-list/components-list.component';
 
 @NgModule({
   imports: [
@@ -27,21 +27,18 @@ import { NbLayoutThemeToggleComponent } from './layout-theme-toggle/layout-theme
     RouterModule.forRoot([
       {
         path: '',
-        loadChildren: '../playground/playground.module#NbPlaygroundModule',
+        loadChildren: '../playground/playground.module#PlaygroundModule',
       },
     ], { useHash: true }),
     NbThemeModule.forRoot(),
-    NbPlaygroundSharedModule,
   ],
   declarations: [
-    NbAppComponent,
-    NbLayoutDirectionToggleComponent,
-    NbLayoutThemeToggleComponent,
+    AppComponent,
+    LayoutDirectionToggleComponent,
+    LayoutThemeToggleComponent,
+    ComponentsOverlayComponent,
+    ComponentsListComponent,
   ],
-  entryComponents: [
-    NbDynamicToAddComponent,
-  ],
-  bootstrap: [NbAppComponent],
+  bootstrap: [ AppComponent ],
 })
-export class NbAppModule {
-}
+export class AppModule {}