+++ src/framework/theme/components/tooltip/tooltip.directive.ts
@@ -4,7 +4,16 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-import { AfterViewInit, ComponentRef, Directive, ElementRef, Inject, Input, OnDestroy } from '@angular/core';
+import {
+  AfterViewInit,
+  ComponentFactoryResolver,
+  ComponentRef,
+  Directive,
+  ElementRef,
+  Inject,
+  Input,
+  OnDestroy,
+} from '@angular/core';
 import { takeWhile } from 'rxjs/operators';
 
 import {
@@ -100,25 +109,17 @@ export class NbTooltipDirective implements AfterViewInit, OnDestroy {
   protected ref: NbOverlayRef;
   protected container: ComponentRef<any>;
   protected positionStrategy: NbAdjustableConnectedPositionStrategy;
-  protected triggerStrategy: NbTriggerStrategy;
   protected alive: boolean = true;
 
   constructor(@Inject(NB_DOCUMENT) protected document,
               private hostRef: ElementRef,
               private positionBuilder: NbPositionBuilderService,
-              private overlay: NbOverlayService) {
+              private overlay: NbOverlayService,
+              private componentFactoryResolver: ComponentFactoryResolver) {
   }
 
   ngAfterViewInit() {
-    this.positionStrategy = this.createPositionStrategy();
-    this.ref = this.overlay.create({
-      positionStrategy: this.positionStrategy,
-      scrollStrategy: this.overlay.scrollStrategies.reposition(),
-    });
-    this.triggerStrategy = this.createTriggerStrategy();
-
     this.subscribeOnTriggers();
-    this.subscribeOnPositionChange();
   }
 
   ngOnDestroy() {
@@ -126,11 +127,11 @@ export class NbTooltipDirective implements AfterViewInit, OnDestroy {
   }
 
   show() {
-    this.container = createContainer(this.ref, NbTooltipComponent, {
-      position: this.position,
-      content: this.content,
-      context: this.context,
-    });
+    if (!this.ref) {
+      this.createOverlay();
+    }
+
+    this.openTooltip();
   }
 
   hide() {
@@ -143,6 +144,24 @@ export class NbTooltipDirective implements AfterViewInit, OnDestroy {
     } else {
       this.show();
     }
+  }
+
+  protected createOverlay() {
+    this.positionStrategy = this.createPositionStrategy();
+    this.ref = this.overlay.create({
+      positionStrategy: this.positionStrategy,
+      scrollStrategy: this.overlay.scrollStrategies.reposition(),
+    });
+    this.subscribeOnPositionChange();
+  }
+
+  protected openTooltip() {
+    this.container = createContainer(this.ref, NbTooltipComponent, {
+      position: this.position,
+      content: this.content,
+      context: this.context,
+      cfr: this.componentFactoryResolver,
+    }, this.componentFactoryResolver);
   }
 
   protected createPositionStrategy(): NbAdjustableConnectedPositionStrategy {
@@ -169,7 +188,8 @@ export class NbTooltipDirective implements AfterViewInit, OnDestroy {
   }
 
   protected subscribeOnTriggers() {
-    this.triggerStrategy.show$.pipe(takeWhile(() => this.alive)).subscribe(() => this.show());
-    this.triggerStrategy.hide$.pipe(takeWhile(() => this.alive)).subscribe(() => this.hide());
+    const triggerStrategy = this.createTriggerStrategy();
+    triggerStrategy.show$.pipe(takeWhile(() => this.alive)).subscribe(() => this.show());
+    triggerStrategy.hide$.pipe(takeWhile(() => this.alive)).subscribe(() => this.hide());
   }
 }