+++ src/playground/with-layout/flip-card/flip-card-colors.component.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright Akveo. All Rights Reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'nb-flip-card-colors',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  templateUrl: './flip-card-colors.component.html',
+})
+export class FlipCardColorsComponent {
+}