+++ src/playground/with-layout/progress-bar/progress-bar-showcase.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'nb-progress-bar-showcase',
+  templateUrl: './progress-bar-showcase.component.html',
+})
+
+export class ProgressBarShowcaseComponent {
+}